+++ backend/pokerogue/src/main/java/com/pokerogue/helper/biome/data/Trainer.java
@@ -1,4 +1,4 @@
-package com.pokerogue.helper.biome.domain;
+package com.pokerogue.helper.biome.data;
 
 import java.util.List;
 import lombok.Getter;
